Nova atualização (#7)
Funcionalidades Implementadas

Adicionado suporte à Playlist

Implementado recurso Up Next

Aplicação de Skin personalizada durante a busca por fontes

+++ resources/lib/resolveurl.py
@@ -1,15 +1,26 @@
+# -*- coding: utf-8 -*-
+
 import xbmc
 import xbmcgui
 import xbmcvfs
+import xbmcaddon
 import os
 import sys
 import urllib.request
 import zipfile
 import shutil
 
+try:
+    _addon = xbmcaddon.Addon()
+except RuntimeError:
+    _addon = xbmcaddon.Addon('plugin.video.thethunder')
+
+def getString(string_id):
+    return _addon.getLocalizedString(string_id)
+
 def download_and_install(url, name):
     dialog = xbmcgui.DialogProgress()
-    dialog.create('TheThunder', 'Baixando atualização do ' + name + '...')
+    dialog.create(getString(30808), '{} {}...'.format(getString(30802), name))
 
     addon_path = xbmcvfs.translatePath('special://home/addons/')
     temp_zip = os.path.join(addon_path, 'temp_resolveurl.zip')
@@ -19,7 +30,7 @@ def download_and_install(url, name):
 
     try:
         urllib.request.urlretrieve(url, temp_zip)
-        dialog.update(50, 'Extraindo arquivos...')
+        dialog.update(50, getString(30803))
 
         if os.path.exists(correct_path):
             shutil.rmtree(correct_path)
@@ -33,15 +44,18 @@ def download_and_install(url, name):
         if os.path.exists(wrong_path):
             os.rename(wrong_path, correct_path)
 
-        os.remove(temp_zip)
-        dialog.update(100, 'Atualização concluída!')
+        if os.path.exists(temp_zip):
+            os.remove(temp_zip)
+
+        dialog.update(100, getString(30804))
+
         xbmcgui.Dialog().ok(
-            'TheThunder',
-            'ResolveURL atualizado com sucesso!'
+            getString(30808),
+            getString(30807)
         )
 
     except Exception as e:
-        xbmcgui.Dialog().ok('Erro', 'Falha ao atualizar ResolveURL:\n' + str(e))
+        xbmcgui.Dialog().ok(getString(30805), '{}:\n{}'.format(getString(30806), str(e)))
 
     finally:
         dialog.close()
@@ -51,7 +65,7 @@ def update():
     download_and_install(url, 'ResolveURL')
 
 if __name__ == '__main__':
-    if len(sys.argv) > 1:
-        action = sys.argv[1]
-        if action == 'update':
-            update()
+    if len(sys.argv) > 1 and sys.argv[1] == 'update':
+        update()
+    else:
+        update()
